move components to entrypoint
