+++ pp/entrypoint/bridge/common.cpp
