+++ pp/entrypoint/bridge/common.h
