+++ pp/entrypoint/bridge/head_wal.h
