+++ pp/entrypoint/bridge/series_data_encoder.cpp
@@ -1,6 +1,6 @@
 #include "series_data_encoder.h"
 
-#include "entrypoint_types/encoder.h"
+#include "entrypoint/types/encoder.h"
 #include "primitives/primitives.h"
 #include "prometheus/relabeler.h"
 #include "series_data/data_storage.h"
