+++ pp/entrypoint/bridge/series_data_encoder.h
