+++ pp/entrypoint/bridge/series_data_serialization_serialized_data.h
