+++ pp/entrypoint/bridge/wal_encoder.cpp
@@ -2,8 +2,8 @@
 
 #include <variant>
 
-#include "entrypoint_types/exception.h"
-#include "entrypoint_types/hashdex.h"
+#include "entrypoint/types/exception.h"
+#include "entrypoint/types/hashdex.h"
 #include "primitives/go_slice.h"
 #include "wal/encoder.h"
 #include "wal/wal.h"
