+++ pp/entrypoint/bridge/wal_encoder.h
