+++ pp/entrypoint/types/exception.h
