+++ pp/entrypoint/types/lss.h
